Copy the new AEK key after generating the ACK message
Was sending the K2S ACK using the new key, not the old.  Sensor
is expecting the old key until they see this ACK, so it was being
dropped.

Also, sensor once in K2S sent state, waiting for ACK of new key, must
stay there until the ACK comes because the GW may have already seen
the K2S message and sent an ACK which we missed.

The remaining portions need appropriate timeouts for rekey success.
Should probably resend the initial REKEY request if we timeout at
any stage, but not yet implemented.