Move clock out to separate file, adjust for battery limits
Clock routines pushed to another file.  Also set the max clock
as a function of the battery voltage.  Need to set fuses for
INTOSC + DIV/8 to get a 1MHz startup and then the code will
check battery levels and adjust the max clock to 8 or 4MHz.
Every time the battery/temp sensor is checked we can adjust
this value.

Also add -O2 to hopefully speed up some of the crypto routines
but no extensive testing done yet.