Use 4-byte MAC to ID sensor, allow event messages to retry faster
A 4-byte MAC is auto-generated first run to uniquely identify the
sensor node.  This is stored w/a checksum in EEPROM.

When an event causes the controller to wake up before standard
report-in time, allow it to retry sending that event 10x faster
than a standard checkin retry, hopefully allowing events a
quicker response.