JSON config, mutex protect sensor, web update
Use a JSON (libjson-c) file format for the sensors report file.  As
it writes the AES key out it'll need to be protected by chmod/chown
in practice.

Protect access to gSensor[] via a MUTEX so the web UI and other threads
can't stomp on things or get inconsistent information.

Add simple way to remove a sensor from the gateway via the web GUI.