Update random number generator with ADC, timers
Redo the RNG to do additional twists and include ADC noise and appropriate
timer configs.  Timer1 set to freerunning because WDT is not guaranteed to
be exactly 16ms, nor is the timer itself when running from intosc.  So we
just want to free run as the LSB will still be random and evenly distributed.